chore: rename integ tests to e2e tests
Because that's what they are, really...
