+++ garden-service/package.json
@@ -200,8 +200,7 @@
     "dev": "npm link && npm run watch",
     "dist": "npm run build && ./bin/build-pkg.sh",
     "generate-docs": "node ./build/src/bin/generate-docs.js",
-    "integ": "./test/integ/run",
-    "integ-full": "node ./build/test/integ/integ-full.js",
+    "e2e": "./test/e2e/run",
     "prepare": "npm run snyk-protect",
     "snyk-protect": "snyk protect",
     "test": "node_modules/.bin/mocha",
