+++ garden-service/test/e2e/README.md
@@ -1,6 +1,6 @@
-## Integration Tests
+# End-to-end Tests
 
-To run the integration test suite (currently consisting only of the pre-release tests), use `npm run integ-full` from within the `garden-service` folder. This runs `garden-service/bin/integ-full.ts`.
+To run the e2e test suite, use `npm run e2e` from within the `garden-service` folder.
 
 The script performs the following cleanup operations before running the tests:
 
@@ -8,12 +8,13 @@ The script performs the following cleanup operations before running the tests:
 * Delete the `garden-system--metadata` namespace and all namespaces belonging to example projects;
 * And deletes the `.garden` directory for every example project.
 
-The `integ-full` script supports the following options:
+The `e2e` script supports the following options:
 
 * `binPath`: The Garden binary to use for the tests (defaults to the one in the static directory). Useful for testing release binaries.
 * `only`: Runs only the test sequence specified (e.g. `demo-project` or `vote-helm`).
 
 For example:
-```
-npm run integ-full -- --binPath=/some/path/garden-bin --project=tasks
+
+```sh
+npm run e2e -- --binPath=/some/path/garden-bin --project=tasks
 ```
