--- garden-service/test/integ/garden.yml
@@ -1,32 +1,0 @@
-kind: Project
-name: integ-tests
-environments:
-  - name: testing
-    providers:
-      - name: kubernetes
-        context: gke_garden-dev-200012_europe-west1-b_garden-dev-1
-        namespace: integ-tests
-        defaultHostname: dev-1.sys.garden
-        buildMode: cluster-docker
-        setupIngressController: nginx
-
----
-
-kind: Module
-name: integ-tests
-type: exec
-tests:
-  - name: demo-project
-    command: [npm, run, integ-full, --, --project=demo-project, --showlog=true, --env=testing]
-  - name: tasks # Tests for tasks are currently being skipped
-    command: [npm, run, integ-full, --, --project=tasks, --showlog=true, --env=testing]
-  - name: hot-reload # Tests for hot-reload are currently being skipped
-    command: [npm, run, integ-full, --, --project=hot-reload, --showlog=true, --env=testing]
-  - name: project-variables
-    command: [npm, run, integ-full, --, --project=project-variables, --showlog=true, --env=testing]
-  - name: vote-helm
-    command: [npm, run, integ-full, --, --project=vote-helm, --showlog=true, --env=testing]
-  - name: vote
-    command: [npm, run, integ-full, --, --project=vote, --showlog=true, --env=testing]
-  - name: remote-sources
-    command: [npm, run, integ-full, --, --project=remote-sources, --showlog=true, --env=testing]
