+++ garden-service/test/run-garden.ts
@@ -4,7 +4,7 @@ import * as mlog from "mocha-logger"
 import { resolve } from "path"
 import { sleep } from "../src/util/util"
 import { TimeoutError } from "bluebird"
-import { parseLogEntries, searchLog, findTasks, touchFile, parsedArgs } from "./integ-helpers"
+import { parseLogEntries, searchLog, findTasks, touchFile, parsedArgs } from "./e2e-helpers"
 import { JsonLogEntry } from "../src/logger/writers/json-terminal-writer"
 import { ParameterError } from "../src/exceptions"
 import { dedent, deline } from "../src/util/string"
