+++ package.json
@@ -57,7 +57,7 @@
     "clean": "lerna run clean && git clean -X -f && find . -name \".garden\" -type d -prune -exec rm -rf '{}' '+'",
     "fix-format": "tslint -p . --fix && tsfmt -r",
     "generate-docs": "lerna run generate-docs",
-    "integ": "lerna run integ",
+    "e2e": "lerna run e2e",
     "lint": "tslint -p . && tsfmt --verify && gulp check-licenses",
     "test": "lerna run test"
   },
